Split internal plugin code into multiple files
